Added getJavaInt and getJavaFloat methods to NumberObject; Moved all arithmetic-related methods there as well
